CPPs

+++ C00/ex00/megaphone.cpp
@@ -5,11 +5,11 @@ int	main(int argc, char **argv)
 {
 	for (int i = 1; i < argc; i++)
 	{
-		for (int j = 0; j < strlen(argv[i]); j++)
+		for (int j = 0; j < (int)strlen(argv[i]); j++)
 			argv[i][j] = toupper((unsigned char)argv[i][j]);
 	}
 	for (int i = 1; i < argc; i++)
 		std::cout << (argv[i]);
-		std::cout << "\n";
+	std::cout << "\n";
 	return (0);
 }
